Move common ThreadMXBean testcase code to TestUtilities
Extract the differences into methods implemented by subclasses.  Note that
ThreadMXBeanTestCaseCommon.java is essentially ThreadMXBeanTestCase.java with
the differences encapsulated in an abstract method.  Also fixed warnings due to
missing attributes.
